Introduced 'extractors' concept as the basis for app-specific notifications text handling
